Use sequencing when posting jobs
Retain the posting order by using sequencing when posting
jobs to a jobboard (at least in the impl_zookeeper case
this is possible).

Adjust the interface to recommend that jobboard implementations
post in oldest to newest order (we can not easily enforce this
since some implementations will likely not be able to do this
in every scenario).

Bring in a new zake which has sequencing support (since this
is required to support this feature).

Change-Id: I0c475c3c1d16e9bbc9db6ff0ad6b7103c5948a47