Allow the WBE to use a preexisting executor
A worker task executor is a good candidate for reuse since
it maintains worker knowledge that is valuable to be retained
across engine runs (tasks on which workers for example). In order
for it to be reused we need a way for the WBE to be able to
receive and reuse a previously existing executor.

Change-Id: Ia9a8f4c544b74e12e2cbd6bd941945da1111499c