Remove the _clear method and do not reset the job_watcher
The _clear method is only used in one place (in close) so the
benefit of it being a function is not really valueable so move
its logic to the close method itself.

Also fix that the job_watcher does not need to be reset since
it can now survive session lose/disconnection, setting it to
none is not useful anymore and actually creates more watchers
that provide the same data (which is bad).

Change-Id: I243ec0551a341df74f478b0161ffd559a75278fe