Provide the compilation object instead of just a part of it
The compilation result will be able to contain more than just
an execution graph in the near future so we should provide it
back instead of a subcomponent of the compilation result.

Breaking change: removes access to the execution graph property
and replaces it with a compilation property (that will contain
the execution graph and any other compilation related objects).

Change-Id: Ieba63f5ab3a59c38f4aa2a22d9caad412f8fa85d