Rename inject_task_args to inject_atom_args
Since storage injection can target atoms and not
just tasks (just one type of atom) the method naming
is more appropriate when named as atom injection (and
associated variable name changes reflect this as well).

Also add more stringent checking around the names of
atoms that are targeted for injection (ensuring that
the name provided actually exists in the storage unit).

Change-Id: I0ad0178240613fda166ea8fcdc441b37290445f8